fix exception caused by *dword used in *size_t call parameter
The code was writing a size_t size of data to a dword.

+++ contrib/win32/win32compat/socket.c
@@ -2718,7 +2718,7 @@ int WSHELPwrite(int sfd, const char *buf, unsigned int max)
     case SFD_TYPE_CONSOLE:
     {
       //ret = _write(sfd_to_fd(sfd), buf, max);
-	  DWORD dwWritten = 0 ;
+	  size_t dwWritten = 0 ;
 	  ret = WriteToConsole(sfd_to_handle(sfd), buf, max, &dwWritten, 0) ;
 	  ret = max ;
